requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
Flask-SocketIO==5.4.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
platformdirs==4.3.6
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.2.3
virtualenv==20.27.1
Werkzeug==3.1.3
wsproto==1.2.0
